chore: use getStaticProps for component views
